+++ pkg/resources/testdata/TestAcc_GrantPrivilegesToDatabaseRole/OnSchema_ExactlyOneOf/test.tf
@@ -1,0 +1,9 @@
+resource "snowflake_grant_privileges_to_database_role" "test" {
+  database_role_name = "some_database.role_name"
+  privileges         = ["USAGE"]
+
+  on_schema {
+    schema_name             = "some_database.schema_name"
+    all_schemas_in_database = "some_database"
+  }
+}